Fix type mismatch on Android

+++ Classes/FenneX/Core/Utility/PListPersist.cpp
@@ -216,7 +216,7 @@ CCObject* loadObjectFromFile(const char* name, bool resource)
 #if (CC_TARGET_PLATFORM == CC_PLATFORM_ANDROID)
     if(resource)
     {
-        long bufferSize = 0;
+        ssize_t bufferSize = 0;
         //Load file from apk
         charbuffer = CCFileUtils::sharedFileUtils()->getFileData(name,"r", &bufferSize);
     }
